**Commit Summary:** - Added calorie entry and tracking functionality. - Implemented password reset via email for forgotten passwords. - Improved the user interface (UI).
**Details:**
1. **Calorie Tracking:**
   - Users can now enter and track their daily calorie intake.
   - Calorie data is stored and accessible within the application.

2. **Password Reset via Email:**
   - If a user forgets their password, they can request a reset link via email.
   - The reset link allows users to set a new password securely.

3. **UI Improvements:**
   - Enhanced the overall look and feel of the application.
   - Streamlined navigation and improved user experience

+++ HealthMate_UI/Screens/AgeCalc.cs
@@ -122,12 +122,12 @@ namespace HealthMate_UI
             }
             DateTime now = DateTime.Now;
             TimeSpan ageTimeSpan = now - birthDate;
-            int years = (int)Math.Floor(ageTimeSpan.TotalDays / 365.242199);
-            int months = (int)Math.Floor((ageTimeSpan.TotalDays % 365.242199) / 30.4368499);
+            byte years = (byte)Math.Floor(ageTimeSpan.TotalDays / 365.242199);
+            short months = (short)Math.Floor((ageTimeSpan.TotalDays % 365.242199) / 30.4368499);
             int days = (int)Math.Floor(ageTimeSpan.TotalDays % 30.4368499);
 
-            int nextBirthdayMonths = (int)Math.Floor(((years + 1) * 365.24199 - ageTimeSpan.TotalDays) / 30.4368499);
-            int nextBirthdayDays = (int)Math.Ceiling(((years + 1) * 365.24199 - ageTimeSpan.TotalDays) % 30.4368499);
+            byte nextBirthdayMonths = (byte)Math.Floor(((years + 1) * 365.24199 - ageTimeSpan.TotalDays) / 30.4368499);
+            byte nextBirthdayDays = (byte)Math.Ceiling(((years + 1) * 365.24199 - ageTimeSpan.TotalDays) % 30.4368499);
 
             int totalDays = (int)ageTimeSpan.TotalDays;
             int totalHours = (int)ageTimeSpan.TotalHours;
@@ -195,12 +195,12 @@ namespace HealthMate_UI
             }
             DateTime now = DateTime.Now;
             TimeSpan ageTimeSpan = now - birthDate;
-            int years = (int)Math.Floor(ageTimeSpan.TotalDays / 365.242199);
-            int months = (int)Math.Floor((ageTimeSpan.TotalDays % 365.242199) / 30.4368499);
+            byte years = (byte)Math.Floor(ageTimeSpan.TotalDays / 365.242199);
+            short months = (short)Math.Floor((ageTimeSpan.TotalDays % 365.242199) / 30.4368499);
             int days = (int)Math.Floor(ageTimeSpan.TotalDays % 30.4368499);
 
-            int nextBirthdayMonths = (int)Math.Floor(((years + 1) * 365.24199 - ageTimeSpan.TotalDays) / 30.4368499);
-            int nextBirthdayDays = (int)Math.Ceiling(((years + 1) * 365.24199 - ageTimeSpan.TotalDays) % 30.4368499);
+            byte nextBirthdayMonths = (byte)Math.Floor(((years + 1) * 365.24199 - ageTimeSpan.TotalDays) / 30.4368499);
+            byte nextBirthdayDays = (byte)Math.Ceiling(((years + 1) * 365.24199 - ageTimeSpan.TotalDays) % 30.4368499);
 
             int totalDays = (int)ageTimeSpan.TotalDays;
             int totalHours = (int)ageTimeSpan.TotalHours;
